go fields: move into internal
This is internal API and should not be relied on by anyone
