+++ internal/fields/sql_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/samsarahq/thunder/fields"
+	"github.com/samsarahq/thunder/internal/fields"
 	"github.com/samsarahq/thunder/internal/proto"
 	"github.com/stretchr/testify/assert"
 )
